refactor: auth → gateway/auth, extensibility → skills + CI 래칫 확장
A. core/auth/ → core/gateway/auth/ (인증은 게이트웨이 관심사)
B. core/extensibility/ → core/skills/ (스킬/플러그인/리포트 통합)
C. scripts/check_legacy_imports.py: auth + extensibility 래칫 추가
D. 기존 코드 import 일괄 마이그레이션 + 구 경로 proxy 설치

Beck: 응집도 (auth=gateway, extensibility=skills)
OpenClaw: 책임 도메인 기준 조직
Tests: 3079 passed. Lint/Type: 0 errors.
