+++ core/cli/report_renderer.py
@@ -16,7 +16,7 @@ from typing import Any
 from core.cli.ui.console import console
 from core.cli.ui.status import GeodeStatus
 from core.config import settings
-from core.extensibility.reports import ReportFormat, ReportGenerator, ReportTemplate
+from core.skills.reports import ReportFormat, ReportGenerator, ReportTemplate
 
 log = logging.getLogger(__name__)
