+++ core/skills/templates/report_detailed.md
@@ -1,0 +1,42 @@
+# GEODE Analysis — ${ip_name}
+
+> ${timestamp} · ${template_type} report
+
+---
+
+## Score Overview
+
+| Metric | Value |
+| --- | --- |
+| **Final Score** | **${final_score}** / 100 |
+| **Tier** | **${tier}** |
+
+${subscores_section}
+
+${scoring_breakdown_section}
+
+${synthesis_section}
+
+${analyses_section}
+
+${analyst_reasoning_section}
+
+${evaluators_section}
+
+${psm_section}
+
+${decision_tree_section}
+
+${details_section}
+
+${biasbuster_section}
+
+${cross_llm_section}
+
+${rights_risk_section}
+
+${signals_section}
+
+---
+
+*Generated by GEODE · Powered by Claude Opus 4.6 + LangGraph*
