+++ core/skills/templates/report_summary.md
@@ -1,0 +1,20 @@
+# GEODE Analysis — ${ip_name}
+
+> ${timestamp} · ${template_type} report
+
+---
+
+## Score Overview
+
+| Metric | Value |
+| --- | --- |
+| **Final Score** | **${final_score}** / 100 |
+| **Tier** | **${tier}** |
+
+${subscores_section}
+
+${synthesis_section}
+
+---
+
+*Generated by GEODE v0.10.1 · Powered by Claude Opus 4.6 + LangGraph*
